added routes for getting and searching for various quizzes on the /search page
