+++ public/stylesheets/style.css
@@ -77,6 +77,7 @@ body {
 
 .outline-3a {
     width: 25%;
+
 }
 
 .outline-3b {
@@ -85,4 +86,10 @@ body {
 
 .outline-5a {
     width: 50%
+}
+
+/* overflow scroll bar for quiz search list */
+.quiz-list {
+    height: 300px;
+    overflow-y: scroll;
 }
