+++ views/layouts/main.handlebars
@@ -23,7 +23,7 @@
 
     <h6 class="text-center">***these buttons are for navigation between pages while under construction***</h6>
     <a href="/login"><button class="btn btn-warning">login/sign up page</button></a>
-    <a href="/quizsearch"><button class="btn btn-warning">quiz search page</button></a>
+    <a href="/search"><button class="btn btn-warning">quiz search page</button></a>
     <a href="/end"><button class="btn btn-warning">quiz end page</button></a>
     <a href="/"><button class="btn btn-warning">quiz home page</button></a>
     <a href="/create"><button class="btn btn-warning">quiz creation page</button></a>
