+++ views/partials/quiz-info.handlebars
@@ -1,8 +1,8 @@
 <label for="quiz-title">Quiz Title</label>
-                <input type="text" placeholder="Title..." id="quiz-title" name="quiz-title" class="form-input" />
+                <input type="text" placeholder="Title..." name="quiz-title" class="form-input" />
                 </br></br>
                 <label for="quiz-category">Choose a Category</label>
-                <select name="quiz-category" id="quiz-category">
+                <select name="quiz-category">
                     <option value="General">General</option>
                     <option value="Coding">Coding</option>
                 </select>
