+++ views/quiz-search-page.handlebars
@@ -1,21 +1,36 @@
 <div class="flexbox">
-        <section class="outline-3a">
-            <div class="outline-3aa">
-                <h3 class="text-center">Quiz Search</h3>
-                <form action="submit">
-                    {{> quiz-info }}
-                    </br></br>
-                    <input type="submit" value="Search Quizzes" class="btn btn-success">
-                    </form>
-            </div>
-            <div class="outline-3ab">
-                <h3 class="text-center">Quiz Info</h3>
-            </div>
-            <button class="btn btn-success" onclick="quizStart()">Start</button>        
-        </section>
+    <section class="outline-3a">
+        <div class="outline-3aa">
+            <h3 class="text-center">Quiz Search</h3>
+            <form action="submit">
+                {{> quiz-info }}
+                </br></br>
+                <input type="submit" value="Search Quizzes" class="btn btn-success">
+            </form>
+        </div>
+        <div class="outline-3ab">
+            <h3 class="text-center">Quiz Info</h3>
+        </div>
+        <button class="btn btn-success" onclick="quizStart()">Start</button>
+    </section>
 
-        <section class="outline-3b">
-            <h3 class="text-center" id="searchResult">Displaying Search Result: </h3>
-            
-        </section>
-    </div>
+    <section class="outline-3b">
+        {{#if search.searchedfor}}
+        <h3 class="text-center" id="searchResult">Displaying Search Result: {{search.searched}} </h3>
+        {{else}}
+        <h3 class="text-center" id="searchResult">Displaying Search Result: </h3>
+        {{/if}}
+        <ul class="quiz-list list-group">
+            {{#each quizzes as |quiz|}}
+            <li class="list-group-item">
+                <div class="quiz">
+                    <h4 class="title">{{title}}</h5>
+                        <h5 class="category">{{category}}</h6>
+                </div>
+            </li>
+            {{/each}}
+        </ul>
+    </section>
+</div>
+
+<script src="/javascript/search-quiz.js"></script>
